ciq_helpers.py: Add wrapper for run_cve_search to avoid duplication
1. Moved run_cve_search from check_kernel_commits.py to ciq_helpers.py
2. Created a wrapper that parses the output of run_cve_search and return the
cve number.
3. Used the wrapper instead of doing the same thing twice in check_kernel_commits.py

Bonus: This also reduces the level of identation in check_kernel_commits.
